requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
anthropic>=0.25.0
openai>=1.0.0
simpy>=4.0.1
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
plotly>=5.14.0
networkx>=3.1
deap>=1.4.1